move mock files out of commerce-api folder
