+++ packages/template-retail-react-app/app/commerce-api/index.test.js
@@ -23,7 +23,7 @@ import {
     expiredAuthToken,
     ocapiBasketResponse,
     ocapiFaultResponse
-} from './mock-data'
+} from '../mocks/mock-data'
 
 jest.mock('cross-fetch', () => jest.requireActual('jest-fetch-mock'))
