+++ packages/template-retail-react-app/app/components/product-view-modal/index.test.js
@@ -11,7 +11,7 @@ import ProductViewModal from './index'
 import {renderWithProviders} from '../../utils/test-utils'
 import {fireEvent, screen} from '@testing-library/react'
 import {useDisclosure} from '@chakra-ui/react'
-import mockProductDetail from '../../commerce-api/mocks/variant-750518699578M'
+import mockProductDetail from '../../mocks/variant-750518699578M'
 
 const MockComponent = ({updateCart}) => {
     const {isOpen, onOpen, onClose} = useDisclosure()
