+++ packages/template-retail-react-app/app/mocks/empty-basket.js
