+++ packages/template-retail-react-app/app/mocks/mock-data.js
