+++ packages/template-retail-react-app/app/mocks/product-set-winter-lookM.js
