+++ packages/template-retail-react-app/app/mocks/searchResults.js
