+++ packages/template-retail-react-app/app/pages/cart/index.test.js
@@ -17,8 +17,8 @@ import {
     mockCustomerBaskets,
     mockEmptyBasket,
     mockCartVariant
-} from '../../commerce-api/mock-data'
-import mockVariant from '../../commerce-api/mocks/variant-750518699578M'
+} from '../../mocks/mock-data'
+import mockVariant from '../../mocks/variant-750518699578M'
 import {rest} from 'msw'
 
 //TODO: Unskip this test when fetchedToken bugs are fixed. All the tests should pass when customerId is defined!
