+++ packages/template-retail-react-app/app/partials/product-view/index.test.js
@@ -8,8 +8,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {fireEvent, screen, waitFor} from '@testing-library/react'
-import mockProductDetail from '../../commerce-api/mocks/variant-750518699578M'
-import mockProductSet from '../../commerce-api/mocks/product-set-winter-lookM'
+import mockProductDetail from '../../mocks/variant-750518699578M'
+import mockProductSet from '../../mocks/product-set-winter-lookM'
 import ProductView from './index'
 import {renderWithProviders} from '../../utils/test-utils'
 import userEvent from '@testing-library/user-event'
